+++ src/main/java/com/casseautomatiche/gdo/repository/ScontrinoRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 @Repository
@@ -16,7 +17,7 @@ public interface ScontrinoRepository extends JpaRepository<Scontrino, UUID> {
             SELECT
             sum(s.totale)
             FROM scontrino s WHERE to_char(s.data_emissione,'DD/MM/YYYY')  =  to_char(current_date, 'DD/MM/YYYY')""",nativeQuery = true)
-    Float incassoGiornaliero();
+    BigDecimal incassoGiornaliero();
 
 
 }
